scripts/requirements.txt
numpy==1.21.0
pandas==1.3.0
opencv-python==4.5.0
scikit-image==0.18.0
Pillow==8.0.0
ipython==7.0.0
typing-extensions==4.0.0